requirements.txt
azure-ai-vision-imageanalysis==1.0.0
azure-core==1.35.1
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
isodate==0.7.2
python-dotenv==1.1.1
requests==2.32.5
six==1.17.0
typing_extensions==4.15.0
urllib3==2.5.0
